UpdateChecker: Refactor form to have 1 main class and one inherited
